requirements.txt
ansicon==1.89.0
arrow==1.3.0
asgiref==3.8.1
blessed==1.20.0
certifi==2025.1.31
charset-normalizer==3.4.1
Django==4.2.19
django-picklefield==3.2
django-q==1.3.9
djangorestframework==3.15.2
idna==3.10
jinxed==1.3.0
pillow==11.1.0
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
redis==3.5.3
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
types-python-dateutil==2.9.0.20241206
tzdata==2025.1
urllib3==2.3.0
wcwidth==0.2.13
